src/analysis/tiering.py
import pandas as pd
import numpy as np
import logging
from typing import Dict, List, Tuple
from sklearn.cluster import KMeans
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

# Set up logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def identify_player_tiers(df: pd.DataFrame, min_clusters: int = 2, max_clusters: int = 8) -> Dict[str, pd.DataFrame]:
    """
    Identify natural performance tiers for each position using clustering.
    
    Args:
        df: Player performance dataframe
        min_clusters: Minimum number of clusters to consider
        max_clusters: Maximum number of clusters to consider
        
    Returns:
        Dict[str, pd.DataFrame]: Dictionary of dataframes with tier assignments by position
    """
    # TODO: Implement identify_player_tiers function
    logger.info("Placeholder: Identify player tiers")
    return {"QB_tiers": pd.DataFrame()}

# TODO: Implement other tiering and archetype analysis functions
