src/data/data_loader.py
import pandas as pd
import os
import yaml
import logging
from typing import Dict, List, Optional, Tuple

# Set up logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def load_config(config_path: str = "config/config.yaml") -> dict:
    """
    Load configuration from YAML file.
    
    Args:
        config_path: Path to the configuration file
        
    Returns:
        dict: Configuration dictionary
    """
    try:
        with open(config_path, 'r') as file:
            config = yaml.safe_load(file)
        logger.info(f"Configuration loaded from {config_path}")
        return config
    except Exception as e:
        logger.error(f"Error loading configuration: {e}")
        raise

def load_csv_data(file_path: str, validate: bool = True) -> pd.DataFrame:
    """
    Load CSV data into a pandas DataFrame with validation.
    
    Args:
        file_path: Path to the CSV file
        validate: Whether to validate the data after loading
        
    Returns:
        pd.DataFrame: Loaded data
    """
    # TODO: Implement load_csv_data function
    logger.info(f"Placeholder: Load data from {file_path}")
    return pd.DataFrame()

# TODO: Implement other data loading functions
